% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{finautoml-package}
\alias{finautoml}
\alias{finautoml-package}
\title{finautoml: What the Package Does (Title Case)}
\description{
AutoML algorithm for finding alpha on financial markets. The AutoML contains preprocessing and ML modleing steps with ensembles. The model is applied on 3 datasets.
}
\author{
\strong{Maintainer}: Luka Sikic \email{luka.sikic@gmail.com}

Authors:
\itemize{
  \item Mislav Sagovac \email{mislav.sagovac@contentio.biz}
}

}
