% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinAutoML.R
\name{FinAutoML}
\alias{FinAutoML}
\title{FinAutoML}
\usage{
FinAutoML(tasks, workers = 2L, outer_folds = 3L, inner_evals = 50L)
}
\arguments{
\item{tasks}{mlr3 tasks.}

\item{workers}{number of workers in parallle.}

\item{outer_folds}{number of outer CV folds.}

\item{inner_evals}{number of iterations inside every fold.}
}
\description{
Create mlr3 graph as AUTOML for prediction in finance (investing).
}
