% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGausscovF3st.R
\name{mlr_filters_gausscov_f3st}
\alias{mlr_filters_gausscov_f3st}
\alias{FilterGausscovF3st}
\title{Gauss cov Feature Selection f3}
\description{
Extract important features by calling [gausscov::f3st()] in
package \CRANpkg{gausscov}.
}
\seealso{
Other Filter: 
\code{\link{mlr_filters_gausscov_f1st}}
}
\concept{Filter}
\section{Super class}{
\code{\link[mlr3filters:Filter]{mlr3filters::Filter}} -> \code{FilterGausscovF3st}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterGausscovF3st-new}{\code{FilterGausscovF3st$new()}}
\item \href{#method-FilterGausscovF3st-clone}{\code{FilterGausscovF3st$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="calculate"><a href='../../mlr3filters/html/Filter.html#method-Filter-calculate'><code>mlr3filters::Filter$calculate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="format"><a href='../../mlr3filters/html/Filter.html#method-Filter-format'><code>mlr3filters::Filter$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="help"><a href='../../mlr3filters/html/Filter.html#method-Filter-help'><code>mlr3filters::Filter$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3filters" data-topic="Filter" data-id="print"><a href='../../mlr3filters/html/Filter.html#method-Filter-print'><code>mlr3filters::Filter$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGausscovF3st-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGausscovF3st-new}{}}}
\subsection{Method \code{new()}}{
Create a GaussCov object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGausscovF3st$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterGausscovF3st-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterGausscovF3st-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterGausscovF3st$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
