% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinAutoML.R
\name{FinAutoML}
\alias{FinAutoML}
\title{FinAutoML}
\usage{
FinAutoML(
  tasks,
  workers = 1L,
  outer_folds = 3L,
  inner_evals = 50L,
  outer_cv = rsmp("RollingWindowCV", folds = outer_folds, fixed_window = FALSE)
)
}
\arguments{
\item{tasks}{mlr3 tasks.}

\item{workers}{number of workers in parallel.}

\item{outer_folds}{number of outer CV folds.}

\item{inner_evals}{number of iterations inside every fold.}

\item{outer_cv}{outer CV.}

\item{folds}{Folds.}

\item{fixed_window}{Window ficed or not.}
}
\description{
Create mlr3 graph as AUTOML for prediction in finance (investing).
}
